requirements.txt
streamlit>=1.28
langchain>=0.0.217
openai>=1.2
streamlit-feedback
azure-identity
azure-ai-documentintelligence
azure-search-documents
azure-storage-blob
python-dotenv
python-docx
PyPDF2
PyMuPDF
langchain 
langchain-community
Pillow
azure-cosmos==4.2.0
PyMuPDF
tabula-py
tiktoken